+++ src/TelegramBotApi.php
@@ -108,6 +108,7 @@ use Vjik\TelegramBot\Api\Method\UnpinAllForumTopicMessages;
 use Vjik\TelegramBot\Api\Method\UnpinAllGeneralForumTopicMessages;
 use Vjik\TelegramBot\Api\Method\UnpinChatMessage;
 use Vjik\TelegramBot\Api\Method\UpdatingMessage\EditMessageCaption;
+use Vjik\TelegramBot\Api\Method\UpdatingMessage\EditMessageMedia;
 use Vjik\TelegramBot\Api\Method\UpdatingMessage\EditMessageText;
 use Vjik\TelegramBot\Api\Request\TelegramRequestInterface;
 use Vjik\TelegramBot\Api\Client\TelegramClientInterface;
@@ -128,6 +129,7 @@ use Vjik\TelegramBot\Api\Type\ForceReply;
 use Vjik\TelegramBot\Api\Type\ForumTopic;
 use Vjik\TelegramBot\Api\Type\InlineKeyboardMarkup;
 use Vjik\TelegramBot\Api\Type\InputFile;
+use Vjik\TelegramBot\Api\Type\InputMedia;
 use Vjik\TelegramBot\Api\Type\InputMediaAudio;
 use Vjik\TelegramBot\Api\Type\InputMediaDocument;
 use Vjik\TelegramBot\Api\Type\InputMediaPhoto;
@@ -571,6 +573,31 @@ final class TelegramBotApi
                 $parseMode,
                 $captionEntities,
                 $showCaptionAboveMedia,
+                $replyMarkup,
+            )
+        );
+    }
+
+    /**
+     * @see https://core.telegram.org/bots/api#editmessagemedia
+     *
+     * @psalm-suppress MixedInferredReturnType,MixedReturnStatement
+     */
+    public function editMessageMedia(
+        InputMedia $media,
+        ?string $businessConnectionId = null,
+        int|string|null $chatId = null,
+        ?int $messageId = null,
+        ?string $inlineMessageId = null,
+        ?InlineKeyboardMarkup $replyMarkup = null,
+    ): FailResult|Message|true {
+        return $this->send(
+            new EditMessageMedia(
+                $media,
+                $businessConnectionId,
+                $chatId,
+                $messageId,
+                $inlineMessageId,
                 $replyMarkup,
             )
         );